% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fev1Projection.R
\name{fev1_projection1}
\alias{fev1_projection1}
\title{FEV1 Projection Model 1}
\usage{
fev1_projection1(fev1_0, int_effect, tio = "No")
}
\arguments{
\item{fev1_0}{numeric: Baseline FEV1 score (L)}

\item{int_effect}{numeric: Effect of intervention on lung function (L)}

\item{tio}{string: is patient receiving tiotropium, "Yes" or "No"}
}
\description{
This is the basic model with only the baseline
FEV1 score
}
