% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fev1Projection.R
\name{fev1_projection3}
\alias{fev1_projection3}
\title{FEV1 Projection Model 3}
\usage{
fev1_projection3(fev1_0, int_effect = 0, male, smoking, age, weight, height)
}
\arguments{
\item{fev1_0}{numeric: Baseline FEV1 score (L)}

\item{int_effect}{numeric: Effect of intervention on lung function (L)}

\item{male}{1 or 0, 1 = male, 0 = female}

\item{smoking}{1 or 0, 1 = patient smokes, 0 = patient does not smoke}

\item{age}{integer: patient age}

\item{weight}{integer: patient weight in kilograms}

\item{height}{numeric: patient height in meters}
}
\description{
This is the complete model, but without the O'Conner slope
}
