% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{matchParameters}
\alias{matchParameters}
\title{Checks to make sure the parameters all match}
\usage{
matchParameters(dataParameters, parameters)
}
\arguments{
\item{dataParameters}{the parameters in the given data}

\item{parameters}{the parameters required for the model}
}
\value{
TRUE if all the arguments match, else FALSE
}
\description{
Checks to make sure the parameters all match
}
