% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{projectionChart}
\alias{projectionChart}
\title{Plots FEV1 Projection}
\usage{
projectionChart(results)
}
\arguments{
\item{results}{results data frame from FEV1 Projection Model}
}
\value{
patientData with prediction
}
\description{
Plots FEV1 Projection
}
\examples{
patientData = samplePatients[1,]
results <- fev1_projection3(fev1_0 = patientData$fev1_0, 
int_effect = patientData$int_effect,
male = patientData$male, smoking = 1, age = patientData$age, 
weight = patientData$weight, height = patientData$height)
chart <- projectionChart(results)
chart
}
