% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-init.R
\docType{data}
\name{samplePatients}
\alias{samplePatients}
\title{Sample Patient Characteristics Inputs}
\format{
A data frame with 2 rows and 6 variables
}
\description{
A dataset containing sample patient characteristics to run the prediction model
 variables are as follows:
}
\details{
\itemize{
  \item ID. patient ID
  \item male. whether the patient is male (0,1)
  \item age. the age of the patient (40--90)
  \item smoker. whether the patient is currently a smoker (0,1)
  \item FEV1. Baseline forced expiratroy volume in 1 second in L (0--5)
  \item height. in m (1--2.2)
  \item weight. in kg (20--150)
  }
}
\keyword{datasets}
